+++ CHANGES.md
@@ -1,4 +1,8 @@
 ## FIX-PROD
+- Se crea página para mostrar el calendario anual de los cntroles
+    N[cal_controles.php]
+
+## FIX-PROD
 - Se corrige rango de valores de riesgo en gráfico del dashboard y metricas de riesgo
     M[pages/met_riesgos.php]
     M[site.php]